+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ##
 ## Added
 - error status -9 for Convert Service
+- sr-Cyrl-RS empty file templates
 
 ## 1.1.2
 ## Changed