+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Added
 - the formsdataurl parameter to the Callback
 - the model FormData
+- sr-Cyrl-RS empty file templates
 
 ## 1.1.2
 ## Changed