+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+##
+## Added
+- commands for working with forgotten files
 - sr-Cyrl-RS empty file templates
 
 ## 1.1.2