+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+##
+## Added
+- commands for working with forgotten files
+- sr-Cyrl-RS empty file templates
+
 ## 1.1.2
 ## Changed
 - fixed load properties in ConfigurationUtils