+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - commands for working with forgotten files
 - sr-Cyrl-RS empty file templates
 
+## Changed
+- pdf instead docxf
+
 ## 1.1.2
 ## Changed
 - fixed load properties in ConfigurationUtils