+++ CHANGELOG.md
@@ -3,7 +3,11 @@
 ##
 ## Added
 - error status -9 and -10 to ConvertResponse.Error
+- commands for working with forgotten files
 - sr-Cyrl-RS empty file templates
+
+## Changed
+- pdf instead docxf
 
 ## 1.1.2
 ## Changed